pyproject.toml
[project]
name = "python-mpi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kagglehub>=0.3.6",
    "logging>=0.4.9.6",
    "matplotlib>=3.10.0",
    "mpi4py>=4.0.1",
    "numpy>=2.2.1",
    "pandas>=2.2.3",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "ucimlrepo>=0.0.7",
]
